+++ src/Services/Sse.php
@@ -7,7 +7,7 @@ namespace Putyourlightson\Datastar\Services;
 
 use Illuminate\Support\Facades\View;
 use Putyourlightson\Datastar\Helpers\Request;
-use Putyourlightson\Datastar\Models\Signals;
+use Putyourlightson\Datastar\Validation\SignalValidator;
 use starfederation\datastar\events\EventInterface;
 use starfederation\datastar\events\ExecuteScript;
 use starfederation\datastar\events\Location;
@@ -22,14 +22,27 @@ use Throwable;
 class Sse
 {
     /**
-     * The response data.
+     * Whether the response is a streamed response.
      */
-    private string $responseData = '';
+    private bool $isStreamedResponse = false;
 
     /**
-     * Whether to send SSE events when processing them.
+     * Whether the session should be closed when the event stream begins.
+     * This is useful to allow other requests to be processed while the event stream is being sent.
      */
-    private bool $sendSseEvents = true;
+    private bool $shouldCloseSession = true;
+
+    /**
+     * Server sent events to send.
+     *
+     * @var EventInterface[]
+     */
+    private array $sseEvents = [];
+
+    /**
+     * Server sent event options to send.
+     */
+    private array $sseEventOptions = [];
 
     /**
      * The server sent event method currently in process.
@@ -42,17 +55,73 @@ class Sse
     private array|null $sseOptionsInProcess = [];
 
     /**
-     * Returns a streamed response.
+     * Returns an event stream.
      */
-    public function getStreamedResponse(callable $callable): StreamedResponse
+    public function getEventStream(?callable $callable = null): StreamedResponse
     {
-        $response = new StreamedResponse($callable);
+        // Abort the process if the client closes the connection.
+        ignore_user_abort(false);
 
-        foreach (ServerSentEventGenerator::headers() as $name => $value) {
-            $response->headers->set($name, $value);
+        $this->isStreamedResponse = true;
+
+        $eventStream = function() use ($callable) {
+            if ($this->shouldCloseSession && session_status() === PHP_SESSION_ACTIVE) {
+                session_write_close();
+            }
+
+            echo $this->getEventOutput();
+            ob_flush();
+            flush();
+
+            if (is_callable($callable)) {
+                $callable();
+            }
+        };
+
+        return new StreamedResponse($eventStream, 200, ServerSentEventGenerator::headers());
+    }
+
+    /**
+     * Returns the output of all events as a string.
+     */
+    public function getEventOutput(bool $reset = true): string
+    {
+        $data = '';
+        foreach ($this->sseEvents as $event) {
+            $data .= $event->getOutput();
         }
 
-        return $response;
+        if ($reset) {
+            $this->resetEvents();
+        }
+
+        return $data;
+    }
+
+    /**
+     * Returns a validator for the signals passed into the request.
+     */
+    public function getValidator(array $rules, array $messages = [], array $attributes = []): SignalValidator
+    {
+        return Request::getValidator($rules, $messages, $attributes);
+    }
+
+    /**
+     * Validates the signals passed into the request.
+     */
+    public function validate(array $rules, array $messages = [], array $attributes = []): array
+    {
+        return Request::getValidator($rules, $messages, $attributes)
+            ->validate();
+    }
+
+    /**
+     * Validates the signals passed into the request using a provided error bag.
+     */
+    public function validateWithBag(string $errorBag, array $rules, array $messages = [], array $attributes = []): array
+    {
+        return Request::getValidator($rules, $messages, $attributes)
+            ->validateWithBag($errorBag);
     }
 
     /**
@@ -66,77 +135,92 @@ class Sse
     /**
      * Patches elements into the DOM.
      */
-    public function patchElements(string $data, array $options = [], bool $send = true): void
+    public function patchElements(string $data, array $options = []): static
     {
         $options = $this->patchEventOptions(
             config('datastar.defaultElementOptions', []),
+            $this->sseEventOptions,
             $options,
         );
         $event = new PatchElements($data, $options);
 
-        $this->processEvent($event, $send);
+        $this->processEvent($event);
+
+        return $this;
     }
 
     /**
      * Removes elements from the DOM.
      */
-    public function removeElements(string $selector, array $options = [], bool $send = true): void
+    public function removeElements(string $selector, array $options = []): static
     {
         $options = $this->patchEventOptions(
             config('datastar.defaultElementOptions', []),
+            $this->sseEventOptions,
             $options,
         );
         $event = new RemoveElements($selector, $options);
 
-        $this->processEvent($event, $send);
+        $this->processEvent($event);
+
+        return $this;
     }
 
     /**
      * Patches signals.
      */
-    public function patchSignals(array $signals, array $options = [], bool $send = true): void
+    public function patchSignals(array $signals, array $options = []): static
     {
         $options = $this->patchEventOptions(
             config('datastar.defaultSignalOptions', []),
+            $this->sseEventOptions,
             $options,
         );
         $event = new PatchSignals($signals, $options);
 
-        $this->processEvent($event, $send);
+        $this->processEvent($event);
+
+        return $this;
     }
 
     /**
      * Executes JavaScript in the browser.
      */
-    public function executeScript(string $script, array $options = [], bool $send = true): void
+    public function executeScript(string $script, array $options = []): static
     {
         $options = $this->patchEventOptions(
             config('datastar.defaultExecuteScriptOptions', []),
+            $this->sseEventOptions,
             $options,
         );
         $event = new ExecuteScript($script, $options);
 
-        $this->processEvent($event, $send);
+        $this->processEvent($event);
+
+        return $this;
     }
 
     /**
      * Redirects the browser by setting the location to the provided URI.
      */
-    public function location(string $uri, array $options = [], bool $send = true): void
+    public function location(string $uri, array $options = []): static
     {
         $options = $this->patchEventOptions(
             config('datastar.defaultExecuteScriptOptions', []),
+            $this->sseEventOptions,
             $options,
         );
         $event = new Location($uri, $options);
 
-        $this->processEvent($event, $send);
+        $this->processEvent($event);
+
+        return $this;
     }
 
     /**
-     * Returns a rendered Datastar view.
+     * Renders a view.
      */
-    public function renderDatastarView(string $view, array $variables = []): string
+    public function renderView(string $view, array $variables = []): static
     {
         if (!View::exists($view)) {
             $this->throwException('View `' . $view . '` does not exist.');
@@ -160,16 +244,52 @@ class Sse
             $this->throwException($exception);
         }
 
-        return $output;
+        if (trim($output) !== '') {
+            $this->patchElements($output);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Resets the current events.
+     */
+    public function resetEvents(): static
+    {
+        $this->sseEvents = [];
+
+        return $this;
+    }
+
+    /**
+     * Sets server sent event options for the current request.
+     */
+    public function setSseEventOptions(array $options): static
+    {
+        $this->sseEventOptions = $options;
+
+        return $this;
     }
 
     /**
      * Sets the server sent event method and options currently in process.
      */
-    public function setSseInProcess(?string $method, array $options = []): void
+    public function setSseInProcess(?string $method, array $options = []): static
     {
         $this->sseMethodInProcess = $method;
         $this->sseOptionsInProcess = $options;
+
+        return $this;
+    }
+
+    /**
+     * Determines whether the session should be closed when the event stream begins.
+     */
+    public function shouldCloseSession(bool $value): static
+    {
+        $this->shouldCloseSession = $value;
+
+        return $this;
     }
 
     /**
@@ -210,34 +330,25 @@ class Sse
     /**
      * Processes an event.
      */
-    private function processEvent(EventInterface $event, bool $send): void
+    private function processEvent(EventInterface $event): void
     {
         $this->verifySseMethodInProcess($event);
 
-        $shouldSend = $this->sendSseEvents && $send;
+        $this->sseEvents[] = $event;
 
-        if ($shouldSend) {
+        if ($this->isStreamedResponse) {
             // Clean and end all existing output buffers.
             while (ob_get_level() > 0) {
                 ob_end_clean();
             }
-        }
 
-        $output = $event->getOutput();
-
-        if ($shouldSend) {
-            echo $output;
+            echo $event->getOutput();
 
             if (ob_get_contents()) {
                 ob_end_flush();
             }
             flush();
-        }
 
-        // Append the resulting output to the response data.
-        $this->responseData .= $output;
-
-        if ($shouldSend) {
             // Start a new output buffer to capture any subsequent inline content.
             ob_start();
         }